% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-block-cd.R
\name{block.cd}
\alias{block.cd}
\title{A block coordinate descent function}
\usage{
block.cd(w.a, ind.m, ind.old, u, y, Tr, X.a, d.link.fn.obj, it.max = 60,
  eps = 10^-4, trace = F, nbasis.t = c(6, 6, 8), rho.grid = c(0,
  0.25, 0.5), linear.link = FALSE, ortho.constr = TRUE)
}
\arguments{
\item{w.a}{a warm start (initial) value of the current nonzero single index coefficients.}

\item{ind.m}{a value of the index used in defining each coordinate-wise block.}

\item{ind.old}{a vector of the indices corresponding to the current nonzero coefficients.}

\item{u}{a vector of the current singe index variable, u = alpha'X.}

\item{y}{treatment outcomes, a n-by-1 vector.}

\item{Tr}{treatment indicators, a n-by-1 vector; each element represents one of the K available treatment options.}

\item{X.a}{a matrix of the pre-treatment covarates, associated with the current nonzero single index coefficients.}

\item{d.link.fn.obj}{an object of the class \code{d.link.fn.obj} passed from the previous iteration.}

\item{it.max}{an integer value specifying the maximum number of iterations for each coordinate.}

\item{eps}{a value specifying the converge criterion of algorithm.}

\item{trace}{if \code{TRUE}, show the trace of the fitting procedure; the default is \code{FALSE}.}

\item{nbasis.t}{a length K+1 vector; each element specifies the number of B-spline basis funtions for approximating the treatment-specific link function; the last element is for the "main effect" link function; the default is \code{nbasis.t=NULL}, and will be determined depending on the sample size.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters for approximating the link functions.}

\item{linear.link}{if \code{TRUE}, restrict the link functions to be linear functions; the default is \code{FALSE}.}

\item{ortho.constr}{the constraint that separates the interaction effects from the main effect (without this, the interaction effect can be confounded by the main effect); the default is \code{TRUE}.}
}
\value{
a list of information of the optimized single index coefficients including
\item{w.a}{a vector of the optimized single index coefficients associated with the active index set.} \item{u}{a vector of the optimized single index variable.} \item{it}{the number of iterations.}
}
\description{
\code{block.cd} implements the block-coordinate descent algorithm of Radchenko (2015) to optimize the (active) single index coefficients.
}
\seealso{
\code{pred.csim},  \code{fit.csim},  \code{fit.csim.cv}
}
