% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{dataGenerationFn}
\alias{dataGenerationFn}
\title{A dataset simulation function}
\usage{
dataGenerationFn(n = 200, p = 50, w = 1, delta = 1,
  true.alpha = NULL, true.eta = NULL, sigma = 0.4,
  correlationX = 0, sigmaX = 1, sim.seed = NULL, obs = FALSE)
}
\arguments{
\item{n}{sample size.}

\item{p}{dimension of the pretreatment covariates.}

\item{w}{controls the nonliarity of the treatment-specific link function that defines the interaction effect component.
\describe{
\item{\code{w=1}}{linear}
\item{\code{w=2}}{almost linear}
\item{\code{w=3}}{highly nonlinear}
}}

\item{delta}{controls the intensity of the main effect.
\describe{
\item{\code{delta=1}}{moderate main effect}
\item{\code{delta=2}}{big main effect}
}}

\item{true.alpha}{a vector of the true single index coefficients (associated with the interaction effect component); if \code{NULL}, then \code{true.alpha} is set to be \code{(1, 0.5, 0.25, 0.125, 0,...0)}' (only the first 4 elements are nonzero).}

\item{true.eta}{a vector of the true main effect coefficients; if \code{NULL}, then \code{true.eta} is randomly generated in a way that only the first 10 elements are nonzero.}

\item{sigma}{standard deviation of the random noise term.}

\item{correlationX}{correlation among the pretreatment covariates.}

\item{sigmaX}{standard deviation of the pretreatment covariates.}

\item{sim.seed}{a seed number for random data generation}

\item{obs}{if \code{TRUE}, use an observational study setting (X and Tr are correlated)}
}
\value{
\item{y}{a n x 1 vector of treatment outcomes.}
\item{Tr}{a n x 1 vector of treatment indicators.}
\item{X}{a n x p matrix of pretreatment covariates.}
\item{SNR}{the "signal" (interaction effects) to "nuisance" (main effects + noise) variance ratio (SNR) of the dataset.}
\item{true.alpha}{the true single index coefficients.}
\item{true.eta}{the true main effect coefficients.}
\item{optTr}{a n x 1 vector of treatments indicating the optimal treatment selections.}
\item{value.opt}{the "Value" of the optimal treatment selection rule, \code{optTr}.}
}
\description{
\code{dataGenerationFn} generates an example dataset under a model that contains a main effect component, a treatment-by-covariates interaction effect component, and a random noise component.
}
