% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-block-cd.R
\name{deriv.link.fn}
\alias{deriv.link.fn}
\title{A first derivative of the link functions estimator}
\usage{
\method{deriv}{link.fn}(link.fn.obj)
}
\arguments{
\item{link.fn.obj}{a fitted link function object of class \code{link.fn}, obtained from the functions \code{fit.link.fn} or \code{fit.link.fn.gcv}.}
}
\value{
a list of information of the fitted link functions given the current single index variable including
\item{d.link.fn}{a n x 1 vector of the 1st derivatives of the estimated link functions evaluated at the current index u = alpha'X.} \item{d.link.fn1}{an associated n x 1 vector used in computing the gradient with respect to the single index coefficients.} \item{d.link.fn2}{an associated n x 1 vector used in computing the updating rule for the single index coefficients.}
}
\description{
A subfunction to compute the first derivative of the link functions, evaluated at the current single index variable u = alpha'X.
}
\seealso{
\code{fit.link.fn},  \code{fit.link.fn.gcv},  \code{fit.csim}, \code{block.cd}
}
