% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-main.R
\name{fit.csim.cv}
\alias{fit.csim.cv}
\title{A Constrained single index model  (a wrapper function for cross-validation)}
\usage{
fit.csim.cv(y, Tr, X, ortho.constr = TRUE, sparse = TRUE,
  type = "AIC", seed = 1234, lam.by = 0.03, n.lam = 100,
  n.max = 10, eps = 10^-4, it.max = 50, trace = F,
  nbasis.t = NULL, rho.grid = c(0, 0.25, 0.5), eff.aug = FALSE,
  X.aug = NULL, linear.link = FALSE, coef.ini = NULL, i.fx = NULL,
  mc.ini = FALSE, unit.norm = FALSE)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options}

\item{X}{a pretreatment covarate matrix, n-by-p matrix}

\item{ortho.constr}{the constraint that separates the interaction effects from the main effect (without this, the interaction effect can be confounded by the main effect); the default is \code{TRUE}.}

\item{sparse}{if \code{TRUE}, apply L1 regularization when estimating the single index coefficients; the default is \code{TRUE}.}

\item{type}{when  \code{sparse=TRUE}, can choose bewteen \code{"CV"} and \code{"BIC"}, for the sparsity tuninig parameter selection.}

\item{seed}{when  \code{type="CV"}, randomization seed for cross validation.}

\item{lam.by}{a value specifying the grid of the sparsity tuning parameters [1, 1+lam.by, ... 1 + n.lam*lam.by].}

\item{n.lam}{a value specifying the grid of the sparsity tuning parameters [1, 1+lam.by, ... 1 + n.lam*lam.by].}

\item{n.max}{a maximum number of nonzero (active) coeffients in CSIM.}

\item{eps}{a value specifying the converge criterion of algorithm.}

\item{it.max}{an integer value specifying the maximum number of iterations for each coordinate.}

\item{trace}{if \code{TRUE}, show the trace of the fitting procedure; the default is \code{FALSE}.}

\item{nbasis.t}{a length K+1 vector; each element specifies the number of B-spline basis funtions for approximating the treatment-specific link function; the last element is for the "main effect" link function; the default is \code{nbasis.t=NULL}, and will be determined depending on the sample size.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters for approximating the link functions.}

\item{eff.aug}{if \code{TRUE}, perform efficiency augmentation (using a L1 regularized linear regression for the main effects of X); the default is \code{FALSE}.}

\item{X.aug}{a design matrix to be used for efficinecy augmentation; the default is \code{NULL}.}

\item{linear.link}{if \code{TRUE}, restrict the link functions to be linear functions; the default is \code{FALSE}.}

\item{coef.ini}{an initial solution for alpha.coef.}

\item{i.fx}{a value of the index to be fixed throughout the estimation for model identifiability; the default is \code{NULL}, hence it is estimated.}

\item{mc.ini}{if \code{TRUE}, use an estimate obtained from the modified covariate method as an initial solution; this is only applicable for K=2 case; the default is \code{FALSE}.}

\item{unit.norm}{if \code{TRUE}, set the estimated coefficient vector to have a unit norm for model identifiability; the default is to set the component with the largest magnitude to be 1.}
}
\value{
a list of information of the fitted model including
\item{coef.obj}{the fitted \code{fit.csim} object with the optimized tuning parameter.} \item{coef.opt}{the estimated single index coefficients, given the optimized  value of the tuning parameter \code{lam}.} \item{coef.pst.opt}{the re-fitted single index coefficients, given the optimized  value of the tuning parameter \code{lam}; the single index coefficients is re-fitted without regularization given a set of selected covariates implied by the nonzero components of \code{coef}, and store in \code{coef.pst}.} \item{eta.coef}{the  estimated  main effect coefficients if \code{eff.aug=TRUE}.}  \item{intercept.y}{a list of the treatment-specific intercepts of the fitted model} \item{center.X}{a vector whose elements are the means of the pretreatment covariates.} \item{scale.X}{a vector whose elements are the standard deviations of the pretreatment covariates.}
}
\description{
\code{fit.csim.cv} performs a 5 fold cross validation to choose the sparsity tuning parameter, and returns the model with the optimized tuning parameter. If \code{type="AIC"}, then the function returns the model with the lowest value of AIC.
}
\seealso{
\code{csim},  \code{fit.csim}, \code{pred.csim}
}
\author{
Park, Petkova, Tarpey, Ogden
}
