% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-block-cd.R
\name{fit.link.fn.gcv}
\alias{fit.link.fn.gcv}
\title{A wrapper function for the link function estimator}
\usage{
fit.link.fn.gcv(y, Tr, u, nbasis.t = NULL, rho.grid = c(0, 0.25, 0.5),
  linear.link = FALSE, ortho.constr = TRUE, ini = FALSE)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector.}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options.}

\item{u}{a vector of the current singe index variable, u = alpha'X.}

\item{nbasis.t}{a length K+1 vector; each element specifies the number of B-spline basis funtions for approximating the treatment-specific link function; the last element is for the "main effect" link function; the default is \code{nbasis.t=NULL}, and will be determined depending on the sample size.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters for approximating the link functions.}

\item{linear.link}{if \code{TRUE}, restrict the link functions to be linear funtions; the default is \code{FALSE}.}

\item{ortho.constr}{the constraint that separates the interaction effects from the main effect (without this, the interaction effect can be confounded by the main effect); the default is \code{TRUE}.}

\item{ini}{if \code{TRUE}, calculate the MSE using the interaction-effect component only.}
}
\value{
\item{beta.t.coef}{a vector of the estimated treatment-specific B-spline coefficients vectors.} \item{beta.0.coef}{a vector of the estimated B-spline coefficients fitted regardless of the treatment indicators.} \item{smoother}{the \code{smoother} object obtained from \code{smoother.fn} used in fitting the link functions.} \item{resid}{a vector of the residuals from the current fitted constrained single index model.} \item{working.resid}{a vector of the residuals from the current fitted constrained single index model, which accounts for the fitted main effects implied by  \code{beta.0.coef}.} \item{Q}{the criterion value (a MSE estimate) of the current fitted constrained single index model.} \item{rho.opt}{the optimized smoothing parameter used in fitting the link functions.}
}
\description{
This is a wrapper function that fits the (B-spline approximated) link functions, given the current single index u = alpha'X. An optimal smoothing parameter, \code{rho.opt}, is chosen by minimizing the generalized cross validation (GCV) for prediction errors.
}
\seealso{
\code{smoother.fn},  \code{fit.link.fn},  \code{fit.csim}, \code{block.cd}
}
