% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-block-cd.R
\name{fit.link.fn}
\alias{fit.link.fn}
\title{A link function estimator}
\usage{
fit.link.fn(y, smoother, ortho.constr = TRUE, ini = FALSE)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector.}

\item{smoother}{a smoother object obtained from \code{smoother.fn}.}

\item{ortho.constr}{the constraint that separates the interaction effects from the main effect (without this, the interaction effect can be confounded by the main effect); the default is \code{TRUE}.}

\item{ini}{if \code{TRUE}, calculate the MSE using the interaction-effect component only.}
}
\value{
a list of information of the fitted link functions given the current single index variable including
\item{beta.t.coef}{a vector of the estimated treatment-specific B-spline coefficients vectors.} \item{beta.0.coef}{a vector of the estimated B-spline coefficients fitted regardless of the treatment indicators.} \item{smoother}{the \code{smoother} object obtained from \code{smoother.fn} used in fitting the link functions.} \item{resid}{a vector of the residuals from the current fitted constrained single index model.} \item{working.resid}{a vector of the residuals from the current fitted constrained single index model, which accounts for the fitted main effects implied by  \code{beta.0.coef}.} \item{MSE}{the criterion value (a MSE estimate) of the current fitted constrained single index model.} \item{rho.opt}{the optimized smoothing parameter used in fitting the link functions.}
}
\description{
Given a vector of responses \code{y} and a smoother object obtained from \code{smoother.fn}, \code{fit.link.fn} computes the treatment-specific link functions of the constrained single index model.
}
\details{
This is a subfunction that fits the (B-spline approximated) link functions gt, given the current single index u = alpha'X. An optimal smoothing parameter, \code{rho.opt}, is chosen by minimizing the generalized cross validation (GCV) for prediction errors.
}
\seealso{
\code{smoother.fn},  \code{fit.link.fn.gcv},  \code{fit.csim}
}
