% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{K.LR}
\alias{K.LR}
\title{K-separate linear regression models}
\usage{
K.LR(y, Tr, X, use.lasso = TRUE)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options}

\item{X}{a pretreatment covarate matrix, n-by-p matrix}

\item{use.lasso}{if \code{TRUE}, estimate the treatment-specific
linear regression coefficients via the Lasso, with the tuning paramters selected by \code{n.folds} cross-validation for prediction errors.}
}
\value{
a list of information of the K separately fitted linear models.
}
\description{
\code{K.LR} is a function for fitting a system of K separate linear regression models, one for each treatment group, for estimating the treatment-by-covariates interaction effects.
}
\details{
A common way to model the treatment-by-covariates interaction effects is to use a system of separate regression functions;  one for each level of the treatment variable.  For example, one can fit a linear regression model for treatment outcomes, for each treatment group separately. One popular apporoach is to use the Lasso for estimating the regression coefficients, which is implented here.
}
\seealso{
\code{pred.K.LR}
}
