% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{K.SAM}
\alias{K.SAM}
\title{K-separate sparse additive models}
\usage{
K.SAM(y, Tr, X, lambda.opt = NULL, n.folds = 5)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options}

\item{X}{a pretreatment covarate matrix, n-by-p matrix}

\item{lambda.opt}{the sparsity tuning paramters, to be selected by \code{n.folds} cross-validation for prediction errors, if \code{lambda.opt=NULL}; the default is \code{NULL}.}

\item{n.folds}{an integer value specifying the number of folds to be used in cross-validation.}
}
\value{
a list of information of the K separately fitted sparse additive models.
}
\description{
\code{K.SAM} is a function for fitting a system of K separate (sparse) additve models (Ravikumar et al. 2009), one for each treatment group, for estimating the treatment-by-covariates (possibly nonlinear) interaction effects.
}
\details{
A common way to model the treatment-by-covariates interaction effects is to use a system of separate regression functions;  one for each level of the treatment variable.  For example, one can fit a regression model for treatment outcomes, for each treatment group separately. One popular apporoach is to use a sparse additive model for fitting the regression fuunctions, which is implented here.
}
\seealso{
\code{pred.K.SAM}
}
