% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{lgem}
\alias{lgem}
\title{Linear generated effect modifiers method (main)}
\usage{
lgem(y, Tr, X, use.lasso = FALSE, n.folds = 20, eff.aug = FALSE,
  boot.CI = FALSE, n.boot = 200, plots = TRUE)
}
\arguments{
\item{y}{treatment outcomes, a n-by-1 vector}

\item{Tr}{treatment indicators, a n-by-1 vector; each element represents one of the K available treatment options}

\item{X}{pretreatment covarates, a n-by-p matrix}

\item{use.lasso}{if \code{TRUE}, estimate the treatment-specific
linear regression coefficients via the Lasso, with the tuning parameters selected by \code{n.folds} cross-validation for prediction errors; the default is \code{FALSE}.}

\item{n.folds}{an integer value specifying the number of folds to be used in cross-validation.}

\item{eff.aug}{if \code{TRUE}, perform efficiency augmentation (using a L1 regularized linear regression for the main effects of X); the default is \code{FALSE}.}

\item{boot.CI}{if \code{TRUE}, return a \code{boot} object that can be used to construct bootstrap confidence intervals for the single index coefficients; the default is \code{FALSE}.}

\item{n.boot}{when \code{boot.CI=TRUE}, a value specifying the number of bootstrap replications.}

\item{plots}{if \code{TRUE}, produce a plot for the link function estimates and the coefficient estimate.}
}
\value{
a list of information of the fitted model including
\item{alpha.coef}{a vector of the estimated GEM coefficients.} \item{gamma.list}{a list of the treatment-specific slopes} \item{coef.list}{a list of the treatment-specific coefficients, i.e., the slope times the GEM coefficients.}\item{link.fn.plot}{a plot object for the treatment-specific link functions (slopes).} \item{coef.plot}{a plot object for the GEM coefficients.} \item{boot.results}{a \code{boot} object that can be used to construct bootstrap  confidence intervals for the GEM coefficients.}
}
\description{
\code{lgem} is a wrapper function for fitting a generated treatment effect modifiers (GEM) model
by the "Numerator" criterion of Petkova et al. (2016). The GEM variable is defined as a linear combination of pretreatment covariates X, optimized under a linear model framework to exhibit a strong (linear) interaction effect with the treatment indicator.
  CSIM is a semiparametric generalization of the GEM model for modeling possilbly nonlinear interaction effects.
}
\seealso{
\code{fit.lgem}, \code{pred.lgem}
}
