% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{mc}
\alias{mc}
\title{Modified covariate method with efficiency augmentation (main)}
\usage{
mc(y, Tr, X, eff.aug = TRUE, X.aug = NULL, use.lasso = TRUE,
  n.folds = 20, boot.CI = FALSE, n.boot = 500)
}
\arguments{
\item{y}{treatment outcomes, n-by-1 vector}

\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options}

\item{X}{a pretreatment covarate matrix, n-by-p matrix}

\item{eff.aug}{if \code{TRUE}, perform efficiency augmentation (using a L1 regularized linear regression for the main effects of X); the default is \code{FALSE}.}

\item{X.aug}{a design matrix to be used for efficinecy augmentation; the default is \code{NULL}.}

\item{use.lasso}{if \code{TRUE}, estimate the treatment-specific
linear regression coefficients via the Lasso, with the tuning paramters selected by \code{n.folds} cross-validation for prediction errors; the default is \code{FALSE}.}

\item{n.folds}{an integer value specifying the number of folds to be used in cross-validation.}

\item{boot.CI}{if \code{TRUE}, return a \code{boot} object that can be used to construct bootstrap confidence intervals for the coefficients; the default is \code{FALSE}.}

\item{n.boot}{when \code{boot.CI=TRUE}, a value specifying the number of bootstrap replications.}
}
\value{
a list of information of the fitted model including
\item{alpha.coef}{a vector of the estimated interaction effect (contrast) coefficients.} \item{eta.coef}{a vector of the  estimated  main effect coefficients if \code{eff.aug=TRUE}.}  \item{intercept.y}{a list of the treatment-specific intercepts of the fitted model} \item{center.X}{a vector whose elements are the means of the pretreatment covariates.} \item{scale.X}{a vector whose elements are the standard deviations of the pretreatment covariates.} \item{MSE}{the mean squared error criterion value.}
}
\description{
\code{mc} is a function for fitting the modified covariate (linear) model of Tian et al. (2015) to estimate the treatment-by-covariates interactions, with efficiency augmentation. CSIM is a semiparametric generalization of the modified covariate model for modeling possilbly nonlinear interaction effects.
}
\seealso{
\code{pred.mc}
}
