% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{performance_measure}
\alias{performance_measure}
\title{A performance mesure function}
\usage{
performance_measure(pred.test, y, Tr, value.opt = NULL, optTr = NULL)
}
\arguments{
\item{pred.test}{a n x K matrix of predicted values given from any model; each column represents each treatment (trt 1, trt 2,..., trt K)}

\item{y}{a n x 1 vector of treatment outcomes from a testing set}

\item{Tr}{a n x 1  vector of treatment indicators from a testing set}

\item{value.opt}{the optimal value of treatment selection rules for the testing set (only available for a simulated dataset).}

\item{optTr}{a vector of treatments indicating the optimal treatment selections for the testing set (only available for a simulated dataset).}
}
\value{
a list of information contating the estimatd Value and the proportion of correct decisions.
}
\description{
\code{performance_measure} assesses the performance of a given treatment selection rule,
in terms of the "Value" of the treatment selection rule and the proportion of correctly made decisions (PCD).
}
\seealso{
\code{dataGenerationFn}
}
