% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-main.R
\name{pred.csim}
\alias{pred.csim}
\title{\code{csim} prediction function}
\usage{
pred.csim(csim.obj, newx)
}
\arguments{
\item{csim.obj}{a \code{csim} object}

\item{newx}{a matrix of new values for the pretreatment covariates at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
This function makes predictions from the constrained single index model
given a \code{csim} object and pretreatment covariates.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{csim},\code{fit.csim}, \code{fit.csim.cv}
}
\author{
Park, Petkova, Tarpey, Ogden
}
