% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{pred.K.LR}
\alias{pred.K.LR}
\title{\code{K.LR} prediction function}
\usage{
pred.K.LR(K.LR.obj, newx)
}
\arguments{
\item{K.LR.obj}{a \code{K.LR} object}

\item{newx}{a matrix of new values for the pretreatment covariates at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
\code{pred.K.LR} makes predictions from the fitted K separate linear models
given a \code{K.LR} object and a set of pretreatment covariates.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{K.LR}
}
