% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{pred.K.SAM}
\alias{pred.K.SAM}
\title{\code{K.SAM} prediction function}
\usage{
pred.K.SAM(K.SAM.obj, newx)
}
\arguments{
\item{K.SAM.obj}{a \code{K.SAM} object}

\item{newx}{a matrix of new values for the pretreatment covariates at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
\code{pred.K.SAM} makes predictions from the fitted K separate sparse additive models
given a \code{K.SAM} object and a set of pretreatment covariates.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{K.SAM}
}
