% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{pred.lgem}
\alias{pred.lgem}
\title{\code{lgem} prediction function}
\usage{
pred.lgem(lgem.obj, newx)
}
\arguments{
\item{lgem.obj}{a \code{lgem} object}

\item{newx}{a matrix of new values for the pretreatment covariates at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
\code{pred.lgem} makes predictions from the linear GEM model
given a \code{lgem} object and a set of pretreatment covariates.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{lgem},\code{fit.lgem}
}
