% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-other-functions.R
\name{pred.mc}
\alias{pred.mc}
\title{\code{mc} prediction function}
\usage{
pred.mc(mc.obj, newx)
}
\arguments{
\item{mc.obj}{a \code{mc} object}

\item{newx}{a matrix of new values for the pretreatment covariates at which predictions are to be made.}
}
\value{
\item{pred.new}{a matrix of predicted values; column represents each treatment.}
\item{trt.rule}{a vector of suggested treatments}
}
\description{
\code{pred.mc} makes predictions from the modfiied covariate model
given a \code{mc} object and a set of pretreatment covariates.
The function returns predicted outcomes for each treatment and treatment selection rules.
}
\seealso{
\code{mc}
}
