% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim-block-cd.R
\name{smoother.fn}
\alias{smoother.fn}
\title{A smoother constructor function}
\usage{
smoother.fn(Tr, u, nbasis.t = c(6, 6, 8), rho.grid = c(0, 0.25, 0.5),
  linear.link = FALSE)
}
\arguments{
\item{Tr}{treatment indicators, n-by-1 vector; each element represents one of the K available treatment options.}

\item{u}{a vector of the current singe index variable, u = alpha'X.}

\item{nbasis.t}{a length K+1 vector; each element specifies the number of B-spline basis funtions used in approximating the treatment-specific link function; the last element is for the "main effect" link function.}

\item{rho.grid}{a grid vector of (ridge-type) smoothing parameters for approximating the link functions.}

\item{linear.link}{if \code{TRUE}, restrict the link functions to be linear funtions; the default is \code{FALSE}.}
}
\value{
a list of information of the smoother matrices given the current single index variable including
\item{Bt.qr}{the QR decomposed design matrix constructed based on both the single index variable and the treatment indicator.} \item{B0.qr}{the QR decomposed design matrix  based on the single index variable only.} \item{Bt}{the design matrix based on both the single index variable and the treatment indicator.} \item{B0}{the design matrix  based on the single index variable only.}\item{knots.t}{the knot sequences used in constructing the treatment-specific B-spline design matrices.}
}
\description{
Given both a vector of the current single index variable u = alpha'X and a vector of the treatment indicators \code{Tr}, \code{smoother.fn} constructs a set of B-spline  smoother matrices, used in estimating the treatment-specific link functions (see \code{fit.link.fn}).
}
\details{
The function returns a set of the QR decomposed design matrices (and the knot sequences used in constructing the B-spline design matrices), over the values of the (ridge-type)  smoothing parameters, \code{rho.grid}. Since the ridge-type smoothing is  equivalent to a regular least squares estimation with added observations, some psedo observations are added to the design matrices for the case of the nonzero values of \code{rho.grid}.
}
\seealso{
\code{fit.link.fn},  \code{fit.link.fn.gcv},  \code{fit.csim}
}
